поправил все ошибки

+++ src/components/Card/UserCards/UserCards.tsx
@@ -9,27 +9,31 @@ import {
 	getWorkoutsById,
 } from "../../../utils/api";
 import { useUser } from "../../../hooks/useUser";
+import { TrainingType } from "../../../types/training";
 
 type UserCardsProps = CardType & { onDelete: (courseId: string) => void };
 
 function UserCards({ courseId, image, nameRu, onDelete }: UserCardsProps) {
 	const [isTrainingSelectModalOpen, setTrainingSelectModalOpen] = useState(false);
-	const [courseData, setCourseData] = useState<any[]>([]);
-	const [workoutInfo, setWorkoutInfo] = useState<any[]>([]);
-	const [completeArray, setCompleteArray] = useState<any[]>([]);
+	const [courseData, setCourseData] = useState([]);
+	const [workoutInfo, setWorkoutInfo] = useState<TrainingType[]>([]);
+	const [completeArray, setCompleteArray] = useState<TrainingType[]>([]);
 	const [isLoading, setIsLoading] = useState(false);
 	const { user } = useUser();
+	const FULL_PROGRESS = 100;
 
 	const openTrainingSelectModal = () => setTrainingSelectModalOpen(true);
 	const closeTrainingSelectModal = () => setTrainingSelectModalOpen(false);
 
 	useEffect(() => {
-		getCourseById(courseId)
-			.then((data) => {
-				setCourseData(data.workouts);
-			})
-			.catch((error) => console.error(error));
-	}, [courseId]);
+		if (user && user.uid) {
+			getCourseById(courseId)
+				.then((data) => {
+					setCourseData(data.workouts);
+				})
+				.catch((error: unknown) => console.error(error));
+		}
+	}, [courseId, user]);
 
 	useEffect(() => {
 		const fetchWorkoutInfo = async () => {
@@ -42,7 +46,7 @@ function UserCards({ courseId, image, nameRu, onDelete }: UserCardsProps) {
 						}),
 					);
 					setWorkoutInfo(workoutInfoArray);
-				} catch (error) {
+				} catch (error: unknown) {
 					console.error("Ошибка при получении информации о курсе:", error);
 				}
 			}
@@ -52,7 +56,7 @@ function UserCards({ courseId, image, nameRu, onDelete }: UserCardsProps) {
 
 	useEffect(() => {
 		const fetchCompleteData = async () => {
-			if (workoutInfo.length > 0) {
+			if (workoutInfo.length > 0 && user && user.uid) {
 				const trainingArray = await Promise.all(
 					workoutInfo.map(async (workout) => {
 						const data = await getRealQuantityWithoutExercises(user.uid, courseId, workout._id);
@@ -64,22 +68,26 @@ function UserCards({ courseId, image, nameRu, onDelete }: UserCardsProps) {
 		};
 		fetchCompleteData();
 		setIsLoading(true);
-	}, [workoutInfo, user.uid, courseId]);
+	}, [workoutInfo, user, courseId]);
 
 	function deleteCourse() {
-		deleteCourseToUser(user.uid, courseId)
-			.then(() => {
-				onDelete(courseId);
-			})
-			.catch((error) => {
-				console.error("Ошибка при удалении курса:", error);
-			});
+		if (user && user.uid) {
+			deleteCourseToUser(user.uid, courseId)
+				.then(() => {
+					onDelete(courseId);
+				})
+				.catch((error: unknown) => {
+					console.error("Ошибка при удалении курса:", error);
+				});
+		}
 	}
 
 	function restartCourse() {
-		deleteProgress(user.uid, courseId).then(() => {
-			setCompleteArray([]);
-		});
+		if (user && user.uid) {
+			deleteProgress(user.uid, courseId).then(() => {
+				setCompleteArray([]);
+			});
+		}
 	}
 
 	const visitedRatio = workoutInfo.length > 0 ? (completeArray.length / workoutInfo.length) * 100 : 0;
@@ -120,10 +128,7 @@ function UserCards({ courseId, image, nameRu, onDelete }: UserCardsProps) {
 								</svg>
 								20-50 мин/день
 							</p>
-							<p
-								className="param
-eter bg-[#F7F7F7] p-2.5 rounded-full flex flex-row gap-1.5 items-center"
-							>
+							<p className="parameter bg-[#F7F7F7] p-2.5 rounded-full flex flex-row gap-1.5 items-center">
 								<svg className="w-[18px] h-[18px]">
 									<use xlinkHref="./icon/sprite.svg#icon-complexity" />
 								</svg>
@@ -139,7 +144,7 @@ eter bg-[#F7F7F7] p-2.5 rounded-full flex flex-row gap-1.5 items-center"
 								</div>
 							</div>
 							<div className="items-center mt-[40px]">
-								{Math.round(visitedRatio) !== 100 ? (
+								{Math.round(visitedRatio) !== FULL_PROGRESS ? (
 									<button
 										onClick={openTrainingSelectModal}
 										className="w-[300px] h-[52px] bg-[#BCEC30] rounded-[46px] hover:bg-[#C6FF00] active:bg-[#000000] active:text-[#FFFFFF] text-lg"
